+++ extensions/easy-new-file/src/types/preferences.ts
@@ -1,6 +1,12 @@
 export interface Preferences {
-  createAndOpen: boolean;
+  layout: string;
+  itemSize: string;
+  itemInset: string;
+  createdActions: string;
   showDocument: boolean;
   showCode: boolean;
   showScript: boolean;
+
+  fileType: string;
+  saveDirectory: string;
 }