+++ extensions/raycast-wallpaper/src/types/preferences.ts
@@ -1,4 +1,6 @@
 export interface Preferences {
+  layout: string;
+  itemSize: string;
   applyTo: string;
   picturesDirectory: string;
 }