+++ extensions/text-shortcuts/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Text Shortcuts Changelog
 
+## [Add new shortcut layout options] - 2022-06-13
+
+- Add new shortcut layout options: List or Grid
+- Optimize shortcut icons
+- Fix text overflow bug
+
 ## [Fix Bug] - 2022-05-20
 
 - Release of Text Shortcuts version 1.3