+++ extensions/world-clock/CHANGELOG.md
@@ -1,5 +1,16 @@
 # World Clock Changelog
 
+## [Fix 24hour bug] - 2022-06-18
+
+- Fix 24hour bug
+- Add new action for Query World Time: Toggle Details
+
+## [Add feature: adding aliases to time zones] - 2022-06-10
+
+- Add feature: you can set aliases for starred time zones now
+- Add new preference: Hour 24
+- Optimize list loading speed
+
 ## [Initial Version] - 2022-05-26
 
-Query the current time of a region, ip or domain.
+- Query the current time of a region, ip or domain.